+++ capytaine/bem/problems_and_results.py
@@ -233,6 +233,7 @@ class LinearPotentialFlowProblem:
     def _asdict(self):
         return {"body_name": self.body_name,
                 "water_depth": self.water_depth,
+                "free_surface": self.free_surface,
                 "omega": float(self.omega),
                 "encounter_omega": float(self.encounter_omega),
                 "period": float(self.period),
@@ -327,6 +328,9 @@ class LinearPotentialFlowProblem:
     def make_results_container(self):
         return LinearPotentialFlowResult(self)
 
+    def make_failed_results_container(self, *args, **kwargs):
+        return FailedLinearPotentialFlowResult(self, *args, **kwargs)
+
 
 class DiffractionProblem(LinearPotentialFlowProblem):
     """Particular LinearPotentialFlowProblem with boundary conditions
@@ -372,6 +376,9 @@ class DiffractionProblem(LinearPotentialFlowProblem):
 
     def make_results_container(self, *args, **kwargs):
         return DiffractionResult(self, *args, **kwargs)
+
+    def make_failed_results_container(self, *args, **kwargs):
+        return FailedDiffractionResult(self, *args, **kwargs)
 
 
 class RadiationProblem(LinearPotentialFlowProblem):
@@ -450,6 +457,9 @@ class RadiationProblem(LinearPotentialFlowProblem):
     def make_results_container(self, *args, **kwargs):
         return RadiationResult(self, *args, **kwargs)
 
+    def make_failed_results_container(self, *args, **kwargs):
+        return FailedRadiationResult(self, *args, **kwargs)
+
 
 class LinearPotentialFlowResult:
 
@@ -497,6 +507,13 @@ class LinearPotentialFlowResult:
     __rich_repr__ = LinearPotentialFlowProblem.__rich_repr__
 
 
+class FailedLinearPotentialFlowResult(LinearPotentialFlowResult):
+    def __init__(self, problem, exception):
+        LinearPotentialFlowResult.__init__(self, problem)
+        self.forces = {dof: np.nan for dof in self.influenced_dofs}
+        self.exception = exception
+
+
 class DiffractionResult(LinearPotentialFlowResult):
 
     def __init__(self, problem, *args, **kwargs):
@@ -512,8 +529,16 @@ class DiffractionResult(LinearPotentialFlowResult):
         return [dict(**params,
                      influenced_dof=dof,
                      diffraction_force=self.forces[dof],
-                     Froude_Krylov_force=FK[dof])
+                     Froude_Krylov_force=FK[dof],
+                     kind="DiffractionResult")
                 for dof in self.influenced_dofs]
+
+
+class FailedDiffractionResult(DiffractionResult):
+    def __init__(self, problem, exception):
+        DiffractionResult.__init__(self, problem)
+        self.forces = {dof: np.nan for dof in self.influenced_dofs}
+        self.exception = exception
 
 
 class RadiationResult(LinearPotentialFlowResult):
@@ -546,5 +571,13 @@ class RadiationResult(LinearPotentialFlowResult):
         return [dict(params,
                      influenced_dof=dof,
                      added_mass=self.added_mass[dof],
-                     radiation_damping=self.radiation_damping[dof])
+                     radiation_damping=self.radiation_damping[dof],
+                     kind="RadiationResult")
                 for dof in self.influenced_dofs]
+
+
+class FailedRadiationResult(RadiationResult):
+    def __init__(self, problem, exception):
+        RadiationResult.__init__(self, problem)
+        self.forces = {dof: np.nan for dof in self.influenced_dofs}
+        self.exception = exception